+++ src/types/index.ts
@@ -1,3 +1,11 @@
 export interface HerokuIntegrationConfig {
   apiKey: string;
 }
+
+export type HerokuClientError = {
+  statusCode: number;
+  body: {
+    id: string;
+    message: string;
+  };
+};